Split: take away Py3statusWrapper from __init__.py
